Remove disabling linters comment
