+++ src/modules/itemCheck.js
@@ -10,12 +10,3 @@ export default window.checkingItems = (itemIdForCheck) => {
   toDoList[itemIdForCheck - 1].completed = checkStatus;
   localStorage.setItem('localStorageTasks', JSON.stringify(toDoList));
 };
-
-export const clearDone = () => {
-  const afterClearArray = toDoList.filter((eachItem) => eachItem.completed !== true);
-  for (let i = 0; i < afterClearArray.length; i += 1) {
-    afterClearArray[i].index = i + 1;
-  }
-  localStorage.setItem('localStorageTasks', JSON.stringify(afterClearArray));
-  window.location.reload();
-};
